+++ frontend/src/components/pages/Modeler/ModelerSidebar/PropertiesPanel/PropertiesPanelComponents/PPTaskDropdown.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Select } from 'antd';
 import PropTypes from 'prop-types';
 import styles from '../../ModelerSidebar.module.css';
-import { getRpaTask } from '../../../../../utils/sessionStorageUtils/localSsotController/attributes';
+import { getRpaTask } from '../../../../../../utils/sessionStorageUtils/localSsotController/attributes';
 
 const { Option } = Select;
