+++ frontend/src/components/pages/Modeler/ModelerSidebar/PropertiesPanel/PropertiesPanelSections/PPNameSection.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Input, Typography, Tooltip } from 'antd';
 import { InfoCircleOutlined } from '@ant-design/icons';
-import corporateDesign from '../../../../../layout/corporateDesign';
+import corporateDesign from '../../../../../../layout/corporateDesign';
 
 const { Text } = Typography;
