+++ frontend/src/components/pages/RobotInteractionCockpit/RobotInteractionSections/RobotInteractionInputSection.jsx
